create a new ViewController
