+++ best_app_ever/QuestionaireViewController.swift
@@ -1,0 +1,35 @@
+//
+//  QuestionaireViewController.swift
+//  best_app_ever
+//
+//  Created by Jinglin Liu on 11/21/17.
+//  Copyright © 2017 iOS Class. All rights reserved.
+//
+
+import UIKit
+
+class QuestionaireViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destinationViewController.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
